Update appearance for messages that contain just 3 emojis
